old_src/requirements.txt
numpy
pyaudio
RPi.GPIO
aubio
scipy
librosa
rpi_ws281x
adafruit-circuitpython-neopixel